\name{DDMarkerData-package}
\alias{DDMarkerData-package}
\docType{package}
\title{
A Homo Sapiens Annotation Database for Diagnose and Detect Markers in Extracellular Circulating
}
\description{
A database is called HSAD, IFF, it entails the symbols among the genes, the proteins, the micro RNAs, the isoforms and their sequences by a deductive system. \cr
The main dataset of the package is data(EC0), and the more details you can find in \code{EC} ( \code{\link{EC0}} ), it is the superset of the dataset \code{EC from DDMarker} ( \code{\link{ECV1}} ), helpping DDMarker to diagnose and detect the sequence among the micro RNAs, the genes, the proteins, and the isoforms. \cr
The hyperlink dataset of the package is data(MMC0), Minimal Metabolize Circulation, and the more details you can find in \code{\link{MMC0}}, it comes from the package \code{DDMarkerMMC} \cr
}
\details{
\tabular{ll}{
Package: \tab DDMarkerData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-27\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
LazyData: \tab true\cr
}
}
\author{
Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
Qiong Yu (JLU & UGA) \email{yuqiong@uga.edu} \cr 
Huansheng Cao (UGA) \email{hshcao@uga.edu} \cr 
Guoqing Liu (IMUST & UGA) \email{gqliu@uga.edu} \cr 
Wei Du (JLU & UGA) \email{weidu@uga.edu} \cr 
Yan Wang (JLU & UGA) \email{wy6868@hotmail.com} \cr 
Hao Wu (BIT & UGA) \email{wuhao@uga.edu} \cr 
Jingyu Yang (GSUT & UGA) \email{jyyyjy@uga.edu} \cr 
Xiufeng Liu (GZUCM & UGA) \email{xfliu@uga.edu} \cr 
Yuan Guo (SAHCQMU & UGA) \email{yuanguo@uga.edu} \cr 
Ying Xu (JLU & UGA) \email{xyn@bmb.uga.edu} \cr \cr 
Maintainer: Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
}
\references{
citation("DDMarkerData");
\cr 
\code{[1]} Yan Wang, et al. (2009) \emph{DMarker: A Bio-Marker Inference System for Human Diseases based
on Microarray Gene Expression Data} 2009 \cr
\code{[2]} Juan Cui, et al. (2008) \emph{Computational prediction of human proteins that can be secreted into the bloodstream} BIOINFORMATICS, Vol.24 no. 20 2008 pages 2370-2375 \cr
\code{[3]} Jiaxin Wang, et al. (2013) \emph{Computational Prediction of Human Salivary Proteins from Blood Circulation and Application to Diagnostic Biomarker Identification} PLoS ONE, DOI: 10.1371/journal.pone.0080211, 2013 \cr
\code{[4]} S Hong, et al. (2011) \emph{A Computational Method for Prediction of Excretory Proteins and Application to Identification of cancer markers in urine and application to gastric cancer} PLoS ONE,6(2):e16875, 2011 \cr
\code{[5]} \url{http://bioinfosrv1.bmb.uga.edu/DMarker/} \cr
}
\keyword{ DDMarker }
\keyword{ DDMarkerData }
\keyword{ HSAD }
\keyword{ EC }
\keyword{ MMC }
\seealso{
\code{\link{EC0}} \code{\link{MMC0}} \code{\link{EC1}} \cr
}
